fix(plugins): grpc-web, grpc-gateway: TE trailers
Ensure `TE` headers is properly sent to gRPC upstream server in request
generated from Kong. Previously, the call to `kong.service.request.set_headers`
was not taking effect as the `TE` headers cannot be set through normal
OpenResty APIs; this PR ensures it's set in a similar way as the
`:authority` pseudo-header.