fix(certificate): unable to refresh certificate entity with vault reference when initial with an invalid string (#10984)
When set the cert to an invalid string in the vault, and then correct the
string, kong will not start to use the correct certificate even though `kong
vault` command shows the correct certiticate being returned.

This is because the initial invalid string has already been cached in the L2
cache by mlcache, and it will not be updated via the L3 cache callback
afterward.

When mlcache fails during the execution of the `l1_serializer`, it caches the
result retrieved from the L3 cache callback the first time in the L2 cache
permanently. This means that subsequent calls to `cache.get()` will never fetch
data from the L3 cache but will directly retrieve it from the L2 cache.

To avoid this situation, when the `l1_serializer` fails, we no longer allow
mlcache to store the data retrieved from L3 into L2.