chore(tracing): log instead of throwing error
This module could throw errors if certain validations fail.
While this is useful for type validation, it can be risky when the error logic
depends on more complex conditions.

This commit removes the error thrown when a span's end time is earlier than its
start time. Although this shouldn't happen under normal circumstances, the end
time calculation depends on multiple factors, including values from Nginx,
making it not trivial to ensure it is always valid.
In such cases, logging an error is more appropriate than failing the request.