fix(templates): add missing security headers for Admin GUI (#14287)
This pull request adds Content-Security-Policy and Referrer-Policy headers while serving Admin GUI requests.

Cherry-picking PR in EE was created manually. Please refer to the linked PR.

Since the Content-Security-Policy has caused some issues with the Kong Manager before, we introduce a new configuration parameter admin_gui_csp_header which defaults to "off" to control the availability of this header. Users can opt in manually.

As for the Referrer-Policy header, we will serve 'strict-origin-when-cross-origin' which seems to be the default value today:

Note: This is the default policy if no policy is specified, or if the provided value is invalid (see spec revision November 2020). Previously the default was no-referrer-when-downgrade.

MDN reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy#strict-origin-when-cross-origin_2

FTI-4283