fix(plugins/key-auth): only clear key from enabled sources
Even if config.key_in_query is set to false, the query is still
modified by the cleanup code. i.e. "%20" will be replaced by "+" for
any query parameter. This maybe suprising to the user, as the config
option was explicitly disabled.
For key_in_body this is already handled differently. I.e. body is
only modified if key_in_body is true.
To make the cleanup code behave more consistent and expectable this PR
makes the cleanup also conditional for query and header.